Add link to counter
The counter was created back in the spring of 2020. Little did I know that "isolation" would last so long! While Benevity is back to working from the office (in a very modified way), the COVID-19 pandemic continues to divide people and isolate us. At this point, the only thing that I can see helping us return to the "old normal" way of life and socially interacting on a regular basis would be a vaccine.